+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Version 2024.3
 *in development*
 
 * 🔧 Android 15: turn on predictive back animation.
+* 🌟 Shows: add permanent year and language filters to Discover tab.
 * 📝 Latest user interface translations from Crowdin.
 
 #### 2024.3.3 🧪